requirements.txt
sagemaker
scikit-learn
pandas
numpy
fastapi
uvicorn  # ASGI server for FastAPI
gunicorn #WSGI server for FastAPI
tensorflow # or tensorflow-cpu or tensorflow-gpu
joblib
boto3
python-multipart  # Required by FastAPI for file uploads
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-prometheus
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests

pandas
scikit-learn
google-cloud-storage
mlflow
optuna
xgboost
pandera